refactor: restructured the solution
