+++ FridgeManager.Contracts/FridgeManager.Contracts.csproj
@@ -1,0 +1,11 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net5.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\FridgeManager.Entities\FridgeManager.Entities.csproj" />
+  </ItemGroup>
+
+</Project>
