+++ FridgeManager.Contracts/IProductService.cs
@@ -1,4 +1,4 @@
-﻿using FridgeManager.Models;
+﻿using FridgeManager.Entities.Models;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
