+++ FridgeManager.Entities/DataTransferObjects/FridgeProductToCreate.cs
@@ -1,16 +1,13 @@
 ﻿using System;
 using System.ComponentModel.DataAnnotations;
 
-namespace FridgeManager.Models
+namespace FridgeManager.Entities.DataTransferObjects
 {
-    public class FridgeProduct
+    public class FridgeProductToCreate
     {
-        public Guid Id { get; set; }
         [Required]
         [Display(Name = "Product Id")]
         public Guid ProductId { get; set; }
-        [Display(Name = "Product Name")]
-        public string ProductName { get; set; }
         [Required]
         [Range(0, int.MaxValue, ErrorMessage = "Quantity can't be less than 0.")]
         public int Quantity { get; set; }
