+++ FridgeManager.Entities/DataTransferObjects/FridgeToUpdate.cs
@@ -1,17 +1,16 @@
 ﻿using System;
 using System.ComponentModel.DataAnnotations;
 
-namespace FridgeManager.Models
+namespace FridgeManager.Entities.DataTransferObjects
 {
-    public class Fridge
+    public class FridgeToUpdate
     {
-        public Guid Id { get; set; }
         [Required]
         public string Name { get; set; }
         [Display(Name = "Owner Name")]
         public string OwnerName { get; set; }
         [Required]
-        [Display(Name = "Model")]
+        [Display(Name = "Model Id")]
         public Guid ModelId { get; set; }
     }
 }
