+++ FridgeManager.Entities/Models/Fridge.cs
@@ -1,16 +1,17 @@
 ﻿using System;
 using System.ComponentModel.DataAnnotations;
 
-namespace FridgeManager.Models.DataTransferObjects
+namespace FridgeManager.Entities.Models
 {
-    public class FridgeToCreate
+    public class Fridge
     {
+        public Guid Id { get; set; }
         [Required]
         public string Name { get; set; }
         [Display(Name = "Owner Name")]
         public string OwnerName { get; set; }
         [Required]
-        [Display(Name = "Model Id")]
+        [Display(Name = "Model")]
         public Guid ModelId { get; set; }
     }
 }
