+++ FridgeManager.Services/ProductService.cs
@@ -1,6 +1,6 @@
 ﻿using FridgeManager.Contracts;
-using FridgeManager.Models;
-using FridgeManager.Models.DataTransferObjects;
+using FridgeManager.Entities.Models;
+using FridgeManager.Entities.DataTransferObjects;
 using System;
 using System.Collections.Generic;
 using System.Linq;
