+++ FridgeManager.sln
@@ -5,6 +5,12 @@ VisualStudioVersion = 17.2.32526.322
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FridgeManager", "FridgeManager\FridgeManager.csproj", "{BA22FEF1-B1C7-4457-BCF0-B5FE9DB29AAE}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FridgeManager.Entities", "FridgeManager.Entities\FridgeManager.Entities.csproj", "{73146CB1-6465-455A-9367-1AE4D6C9D712}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FridgeManager.Contracts", "FridgeManager.Contracts\FridgeManager.Contracts.csproj", "{04A9E123-E987-433D-87D6-05ECBBFB545C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FridgeManager.Services", "FridgeManager.Services\FridgeManager.Services.csproj", "{A211A982-CB7E-4582-9679-28B8835DCDD4}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -15,6 +21,18 @@ Global
 		{BA22FEF1-B1C7-4457-BCF0-B5FE9DB29AAE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BA22FEF1-B1C7-4457-BCF0-B5FE9DB29AAE}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{BA22FEF1-B1C7-4457-BCF0-B5FE9DB29AAE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{73146CB1-6465-455A-9367-1AE4D6C9D712}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{73146CB1-6465-455A-9367-1AE4D6C9D712}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{73146CB1-6465-455A-9367-1AE4D6C9D712}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{73146CB1-6465-455A-9367-1AE4D6C9D712}.Release|Any CPU.Build.0 = Release|Any CPU
+		{04A9E123-E987-433D-87D6-05ECBBFB545C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{04A9E123-E987-433D-87D6-05ECBBFB545C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{04A9E123-E987-433D-87D6-05ECBBFB545C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{04A9E123-E987-433D-87D6-05ECBBFB545C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A211A982-CB7E-4582-9679-28B8835DCDD4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A211A982-CB7E-4582-9679-28B8835DCDD4}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A211A982-CB7E-4582-9679-28B8835DCDD4}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A211A982-CB7E-4582-9679-28B8835DCDD4}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
