+++ FridgeManager/Models/ReauiredAttribute.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace FridgeManager.Models
+{
+    internal class ReauiredAttribute : Attribute
+    {
+    }
+}
