+++ FridgeManager/Views/Fridges/CreateProduct.cshtml
@@ -1,4 +1,4 @@
-﻿@model FridgeManager.Models.ViewModels.CreateFridgeProductViewModel
+﻿@model FridgeManager.Models.CreateFridgeProductViewModel
 
 @{
     ViewData["Title"] = "Create Product";
