+++ FridgeManager/Views/Fridges/Details.cshtml
@@ -1,4 +1,4 @@
-﻿@model FridgeManager.Models.Fridge
+﻿@model FridgeManager.Entities.Models.Fridge
 
 @{
     ViewData["Title"] = "Details";
