+++ FridgeManager/Views/Fridges/Edit.cshtml
@@ -1,4 +1,4 @@
-﻿@model FridgeManager.Models.Fridge
+﻿@model FridgeManager.Models.UpdateFridgeViewModel
 
 <h1>Edit</h1>
 
@@ -6,29 +6,34 @@
 <hr />
 <div class="row">
     <div class="col-md-4">
-        <form asp-action="Edit" id="form">
+        <form asp-action="Edit" asp-rout-id="@Model.Fridge.Id">
             <div asp-validation-summary="ModelOnly" class="text-danger"></div>
             <div class="form-group">
-                <label asp-for="@Model.Name" class="control-label"></label>
-                <input asp-for="@Model.Name" class="form-control" />
-                <span asp-validation-for="@Model.Name" class="text-danger"></span>
+                <label asp-for="@Model.Fridge.Name" class="control-label"></label>
+                <input asp-for="@Model.Fridge.Name" class="form-control" />
+                <span asp-validation-for="@Model.Fridge.Name" class="text-danger"></span>
             </div>
             <div class="form-group">
-                <label asp-for="@Model.OwnerName" class="control-label"></label>
-                <input asp-for="@Model.OwnerName" class="form-control" />
-                <span asp-validation-for="@Model.OwnerName" class="text-danger"></span>
+                <label asp-for="@Model.Fridge.OwnerName" class="control-label"></label>
+                <input asp-for="@Model.Fridge.OwnerName" class="form-control" />
+                <span asp-validation-for="@Model.Fridge.OwnerName" class="text-danger"></span>
             </div>
             <div class="form-group">
-                <label asp-for="@Model.ModelId" class="control-label"></label>
-                <input asp-for="@Model.ModelId" class="form-control" />
-                <span asp-validation-for="@Model.ModelId" class="text-danger"></span>
+                <label asp-for="@Model.Fridge.ModelId" class="control-label"></label>
+                <select asp-for=@Model.Fridge.ModelId>
+                    @foreach (var item in Model.Models)
+                    {
+                        <option value="@item.Id">@item.Name - @item.Year</option>
+                    }
+                </select>
+                <span asp-validation-for="@Model.Fridge.ModelId" class="text-danger"></span>
             </div>
             <div class="row">
-                <div class="col-md-2">
+                <div class="col-md-4">
                     <input type="submit" value="Update" class="btn btn-primary" />
                 </div>
                 <div class="col-md-2">
-                    <a class="btn btn-secondary" asp-action="Index">Back to List</a>
+                    <a class="btn btn-secondary" asp-action="Index">Back</a>
                 </div>
             </div>
         </form>
