+++ FridgeManager/Views/Fridges/FridgeProducts.cshtml
@@ -1,4 +1,5 @@
-﻿@model FridgeManager.Models.ViewModels.CreateFridgeViewModel
+﻿@model FridgeManager.Models.CreateFridgeViewModel
+
 @if(@Model.FridgeProducts.Count() > 0)
 {
     @for (int i = 0; i < Model.FridgeProducts.Count(); i++)
