+++ FridgeManager/Views/Fridges/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<FridgeManager.Models.Fridge>
+﻿@model IEnumerable<FridgeManager.Entities.Models.Fridge>
 
 @{
     ViewData["Title"] = "Index";
