+++ FridgeManager/Views/Fridges/ProductsList.cshtml
@@ -1,4 +1,6 @@
-﻿@model FridgeManager.Models.ViewModels.ProductsListViewModel
+﻿@using FridgeManager.Entities.Models
+@model FridgeManager.Models.ProductsListViewModel
+
 @{
     ViewData["Title"] = "ProductsList";
 }
