+++ FridgeManager/Views/Home/Privacy.cshtml
@@ -3,4 +3,4 @@
 }
 <h1>@ViewData["Title"]</h1>
 
-<p>Use this page to detail your site's privacy policy.</p>
+<p>This application is too simplistic to take any personal information.</p>
