+++ FridgeManager/Views/Products/Create.cshtml
@@ -1,4 +1,4 @@
-﻿@model FridgeManager.Models.DataTransferObjects.ProductToCreate
+﻿@model FridgeManager.Entities.DataTransferObjects.ProductToCreate
 
 @{
     ViewData["Title"] = "Create";
