+++ FridgeManager/Views/Products/Edit.cshtml
@@ -1,4 +1,4 @@
-﻿@model FridgeManager.Models.Product
+﻿@model FridgeManager.Entities.Models.Product
 
 <h1>Edit</h1>
