+++ FridgeManager/Views/Products/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<FridgeManager.Models.Product>
+﻿@model IEnumerable<FridgeManager.Entities.Models.Product>
 
 @{
     ViewData["Title"] = "Index";
