+++ FridgeManager/Views/Shared/EditorTemplates/_FridgeProduct.cshtml
@@ -1,4 +1,4 @@
-﻿@model FridgeManager.Models.DataTransferObjects.FridgeProductToAdd
+﻿@model FridgeManager.Entities.DataTransferObjects.FridgeProductToAdd
 
 <div class="row">
     <div class="col-md-4">
